Sistema de Autenticação de Usuário

+++ ExchangeCripto/src/model/Coin.java
@@ -1,8 +1,9 @@
 package model;
+import java.math.BigDecimal;
 
 public class Coin {
     private int id;
-    private double quotation;
+    private BigDecimal quotation;
     private double taxVenda;
     private double taxCompra;
 }
